+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Method to give the next prime.
+- Generator of prime numbers
 
 ## [0.1.0] - 2019-04-18
 ### Added